refactor Ui dummy stream code
